Refactor to provide test database from the implementation
