+++ src/persistence/repository.rs
@@ -1,26 +1,15 @@
 use async_trait::async_trait;
 use bigdecimal::BigDecimal;
 use sqlx::postgres::PgPoolOptions;
-use sqlx::PgPool;
+use sqlx::{ConnectOptions, Connection, Executor, PgConnection, PgPool};
+use tracing::log::LevelFilter;
 use uuid::Uuid;
 
 use super::model::OcieItem;
+use super::OcieItemRepository;
 use crate::configuration::DatabaseSettings;
 use crate::error_handling::error_chain_fmt;
 use crate::persistence::model::{LineItemNumber, NationalStockNumber};
-
-#[async_trait]
-pub trait OcieItemRepository: Send + Sync + 'static {
-    type Error;
-    //type Connection;
-    type RecordIdType;
-    fn new(database_settings: &DatabaseSettings) -> Self;
-    async fn get_all(&self) -> Result<Vec<OcieItem>, Self::Error>;
-    async fn get(&self, id: Self::RecordIdType) -> Result<OcieItem, Self::Error>;
-    async fn add(&self, item: OcieItem) -> Result<OcieItem, Self::Error>;
-    async fn update(&self, id: Self::RecordIdType, item: OcieItem)
-        -> Result<OcieItem, Self::Error>;
-}
 
 pub struct PostgresOcieItemRepository {
     pool: PgPool,
@@ -34,6 +23,42 @@ impl PostgresOcieItemRepository {
             .connect_timeout(std::time::Duration::from_secs(2))
             .connect_lazy_with(configuration.with_db())
     }
+
+    async fn get_test_connection_pool(configuration: &DatabaseSettings) -> PgPool {
+        // Create a new database
+        let mut connection = PgConnection::connect_with(&configuration.without_db())
+            .await
+            .expect("Failed to connect to postgres.");
+
+        connection
+            .execute(&*format!(
+                r#"CREATE DATABASE "{}";"#,
+                configuration.database_name
+            ))
+            .await
+            .expect("Failed to create database.");
+
+        // Create a database pool for the web server, specifying that sqlx logs
+        // should be at the `tracing` level.
+        let db_connect_options = configuration
+            .with_db()
+            .log_statements(LevelFilter::Trace)
+            .to_owned();
+
+        let connection_pool = PgPoolOptions::new()
+            .connect_timeout(std::time::Duration::from_secs(2))
+            .connect_with(db_connect_options)
+            .await
+            .expect("Failed to connect to Postgres.");
+
+        // Run database migrations
+        sqlx::migrate!("./migrations")
+            .run(&connection_pool)
+            .await
+            .expect("Failed to migrate the database.");
+
+        connection_pool
+    }
 }
 #[async_trait]
 impl OcieItemRepository for PostgresOcieItemRepository {
@@ -43,10 +68,16 @@ impl OcieItemRepository for PostgresOcieItemRepository {
 
     type RecordIdType = Uuid;
 
-    fn new(database_configuration: &DatabaseSettings) -> Self {
+    async fn new(database_configuration: &DatabaseSettings) -> Self {
         Self {
             // Get a connection pool for the database
             pool: Self::get_connection_pool(database_configuration),
+        }
+    }
+
+    async fn new_test_repository(database_configuration: &DatabaseSettings) -> Self {
+        Self {
+            pool: Self::get_test_connection_pool(database_configuration).await,
         }
     }
 
@@ -159,49 +190,6 @@ impl OcieItemRepository for PostgresOcieItemRepository {
     ) -> Result<OcieItem, Self::Error> {
         todo!()
     }
-}
-
-#[tracing::instrument(name = "Get all OcieItems", skip(conn))]
-pub async fn get_all(conn: &PgPool) -> Result<Vec<OcieItem>, anyhow::Error> {
-    let result = sqlx::query!(
-        r#"SELECT id, nsn, lin, nomenclature, size, unit_of_issue, price
-        FROM ocieitems"#
-    )
-    .fetch_all(conn)
-    .await?
-    .into_iter()
-    .flat_map(|row| {
-        let nsn = match NationalStockNumber::parse(row.nsn) {
-            Ok(nsn) => nsn,
-            Err(e) => {
-                tracing::error!("Error parsing NSN: {:?}", e);
-                return None;
-            }
-        };
-        let lin = match LineItemNumber::parse(row.lin) {
-            Ok(lin) => lin,
-            Err(e) => {
-                tracing::error!("Error parsing LIN: {:?}", e);
-                return None;
-            }
-        };
-        Some(OcieItem {
-            id: row.id,
-            nsn,
-            lin,
-            nomenclature: row.nomenclature,
-            size: row.size,
-            unit_of_issue: row.unit_of_issue,
-            price: row.price,
-        })
-    })
-    .collect::<Vec<OcieItem>>();
-    Ok(result)
-}
-
-pub fn get(_id: i32, _conn: &PgPool) -> Result<OcieItem, sqlx::Error> {
-    //TODO: Run SQLX query to pull OcieItem
-    todo!()
 }
 
 #[derive(thiserror::Error)]
